Wip (#1060)
* backup2go list view fix

For the backup2go module the listing view for the pane “Last
successfull backup” didn’t work properly.

* Added comment for the firmware_escrow script

I tried the example script for the firmware_escrow module and it
failed. You need to change the first octet of the ip range to make it
work. I’ve a extra comment to the README.